phase2/backend/requirements.txt
# requirements.txt (updated with Better Auth dependencies)
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg2-binary
pydantic==2.5.0
PyJWT==2.8.0
cryptography==41.0.7
better-auth==0.0.1b11
